+++ .changeset/split-aws-kms-signer.md
@@ -1,0 +1,5 @@
+---
+'@mysten/aws-kms-signer': minor
+---
+
+Initial release. Sui signer using AWS KMS, previously available via `@mysten/signers/aws`.
