+++ .changeset/split-gcp-kms-signer.md
@@ -1,0 +1,5 @@
+---
+'@mysten/gcp-kms-signer': minor
+---
+
+Initial release. Sui signer using Google Cloud KMS, previously available via `@mysten/signers/gcp`.
