+++ .changeset/split-ledger-signer.md
@@ -1,0 +1,5 @@
+---
+'@mysten/ledger-signer': minor
+---
+
+Initial release. Sui signer for Ledger hardware wallets, previously available via `@mysten/signers/ledger`.
