+++ .changeset/split-webcrypto-signer.md
@@ -1,0 +1,5 @@
+---
+'@mysten/webcrypto-signer': minor
+---
+
+Initial release. Sui signer using the Web Crypto API (P-256 / Secp256r1), previously available via `@mysten/signers/webcrypto`.
