+++ packages/docs/package.json
@@ -39,7 +39,11 @@
 		"@mysten/kiosk": "workspace:*",
 		"@mysten/payment-kit": "workspace:*",
 		"@mysten/seal": "workspace:*",
+		"@mysten/aws-kms-signer": "workspace:*",
+		"@mysten/gcp-kms-signer": "workspace:*",
+		"@mysten/ledger-signer": "workspace:*",
 		"@mysten/signers": "workspace:*",
+		"@mysten/webcrypto-signer": "workspace:*",
 		"@mysten/slush-wallet": "workspace:*",
 		"@mysten/sui": "workspace:*",
 		"@mysten/suins": "workspace:*",
