+++ packages/signers/aws/package.json
@@ -1,0 +1,62 @@
+{
+	"name": "@mysten/aws-kms-signer",
+	"author": "Mysten Labs <build@mystenlabs.com>",
+	"description": "Sui signer using AWS KMS",
+	"homepage": "https://sdk.mystenlabs.com",
+	"version": "0.0.0",
+	"license": "Apache-2.0",
+	"type": "module",
+	"sideEffects": false,
+	"exports": {
+		".": {
+			"types": "./dist/index.d.mts",
+			"import": "./dist/index.mjs",
+			"default": "./dist/index.mjs"
+		}
+	},
+	"files": [
+		"CHANGELOG.md",
+		"LICENSE",
+		"README.md",
+		"dist",
+		"src"
+	],
+	"scripts": {
+		"clean": "rm -rf tsconfig.tsbuildinfo ./dist",
+		"build": "rm -rf dist && tsc --noEmit && tsdown",
+		"prettier:check": "prettier -c --ignore-unknown .",
+		"prettier:fix": "prettier -w --ignore-unknown .",
+		"oxlint:check": "oxlint  .",
+		"oxlint:fix": "oxlint --fix",
+		"lint": "pnpm run oxlint:check && pnpm run prettier:check",
+		"lint:fix": "pnpm run oxlint:fix && pnpm run prettier:fix",
+		"test": "vitest run"
+	},
+	"repository": {
+		"type": "git",
+		"url": "git+https://github.com/MystenLabs/ts-sdks.git"
+	},
+	"bugs": {
+		"url": "https://github.com/mystenlabs/ts-sdks/issues"
+	},
+	"publishConfig": {
+		"access": "public"
+	},
+	"engines": {
+		"node": ">=22"
+	},
+	"dependencies": {
+		"@noble/curves": "^2.0.1",
+		"asn1-ts": "^11.0.5"
+	},
+	"peerDependencies": {
+		"@mysten/sui": "workspace:^"
+	},
+	"devDependencies": {
+		"@mysten/sui": "workspace:^",
+		"@types/node": "^25.0.8",
+		"dotenv": "^17.2.3",
+		"typescript": "^5.9.3",
+		"vitest": "^4.0.17"
+	}
+}
