+++ packages/signers/aws/src/aws-client.ts
@@ -5,8 +5,8 @@ import { Secp256k1PublicKey } from '@mysten/sui/keypairs/secp256k1';
 import { Secp256r1PublicKey } from '@mysten/sui/keypairs/secp256r1';
 import { fromBase64 } from '@mysten/sui/utils';
 
-import { publicKeyFromDER } from '../utils/utils.js';
 import { AwsClient } from './aws4fetch.js';
+import { publicKeyFromDER } from './utils.js';
 
 interface KmsCommands {
 	Sign: {
