+++ packages/signers/aws/src/aws4fetch.ts
