+++ packages/signers/aws/src/utils.ts
