+++ packages/signers/aws/vitest.config.ts
@@ -1,0 +1,19 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+	test: {
+		maxWorkers: 4,
+		hookTimeout: 1000000,
+		testTimeout: 1000000,
+		env: {
+			NODE_ENV: 'test',
+		},
+		setupFiles: ['dotenv/config'],
+	},
+	resolve: {
+		alias: {},
+	},
+});
