+++ packages/signers/gcp/src/gcp-kms-client.ts
@@ -7,7 +7,7 @@ import { Secp256k1PublicKey } from '@mysten/sui/keypairs/secp256k1';
 import { Secp256r1PublicKey } from '@mysten/sui/keypairs/secp256r1';
 import { fromBase64 } from '@mysten/sui/utils';
 
-import { getConcatenatedSignature, publicKeyFromDER } from '../utils/utils.js';
+import { getConcatenatedSignature, publicKeyFromDER } from './utils.js';
 
 /**
  * Configuration options for initializing the GcpKmsSigner.
