+++ packages/signers/gcp/src/index.ts
@@ -1,0 +1,9 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+import type { GcpKmsSignerOptions } from './gcp-kms-client.js';
+import { GcpKmsSigner } from './gcp-kms-client.js';
+
+export { GcpKmsSigner };
+
+export type { GcpKmsSignerOptions };
