+++ packages/signers/gcp/tests/e2e-gcp-kms.test.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 import { beforeAll, describe, expect, it } from 'vitest';
 
-import { GcpKmsSigner } from '../src/gcp/gcp-kms-client.js';
+import { GcpKmsSigner } from '../src/gcp-kms-client.js';
 
 const { E2E_GCP_KMS_TEST_ENABLE } = process.env;
