+++ packages/signers/ledger/.prettierignore
@@ -1,0 +1,2 @@
+CHANGELOG.md
+dist/
