+++ packages/signers/ledger/vitest.config.ts
@@ -1,0 +1,12 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+	test: {
+		env: {
+			NODE_ENV: 'test',
+		},
+	},
+});
