+++ packages/signers/webcrypto/tsconfig.json
@@ -1,0 +1,7 @@
+{
+	"extends": "../../../tsconfig.shared.json",
+	"include": ["src"],
+	"compilerOptions": {
+		"noEmit": true
+	}
+}
