+++ pnpm-workspace.yaml
@@ -28,14 +28,9 @@ packages:
   - '!packages/sui/keypairs/secp256r1'
   - '!packages/sui/graphql/schemas/2024.1'
   - '!packages/sui/graphql/schemas/2024.4'
-  - '!packages/signers/aws'
-  - '!packages/signers/gcp'
-  - '!packages/signers/ledger'
   - '!packages/sui/zklogin'
   - '!packages/sui/keypairs/passkey'
   - '!packages/sui/graphql/schema'
-  - '!packages/signers/utils'
-  - '!packages/signers/webcrypto'
   - '!packages/zksend/experimental/unified-adapter'
   - '!packages/sui/experimental'
   - '!packages/dapp-kit-next/packages/dapp-kit-core/web'
